Make selector private and update docs
