+++ packages/block-editor/src/store/test/private-selectors.js
@@ -1,7 +1,10 @@
 /**
  * Internal dependencies
  */
-import { isBlockInterfaceHidden } from '../private-selectors';
+import {
+	isBlockInterfaceHidden,
+	getLastInsertedBlocksClientIds,
+} from '../private-selectors';
 
 describe( 'private selectors', () => {
 	describe( 'isBlockInterfaceHidden', () => {
@@ -19,6 +22,31 @@ describe( 'private selectors', () => {
 			};
 
 			expect( isBlockInterfaceHidden( state ) ).toBe( false );
+		} );
+	} );
+
+	describe( 'getLastInsertedBlocksClientIds', () => {
+		it( 'should return undefined if no blocks have been inserted', () => {
+			const state = {
+				lastBlockInserted: {},
+			};
+
+			expect( getLastInsertedBlocksClientIds( state ) ).toEqual(
+				undefined
+			);
+		} );
+
+		it( 'should return clientIds if blocks have been inserted', () => {
+			const state = {
+				lastBlockInserted: {
+					clientIds: [ '123456', '78910' ],
+				},
+			};
+
+			expect( getLastInsertedBlocksClientIds( state ) ).toEqual( [
+				'123456',
+				'78910',
+			] );
 		} );
 	} );
 } );
