+++ packages/block-editor/src/store/test/selectors.js
@@ -72,7 +72,6 @@ const {
 	__experimentalGetPatternTransformItems,
 	wasBlockJustInserted,
 	__experimentalGetGlobalBlocksByName,
-	getLastInsertedBlocksClientIds,
 } = selectors;
 
 describe( 'selectors', () => {
@@ -4685,29 +4684,6 @@ describe( '__unstableGetClientIdsTree', () => {
 					{ clientId: 'baz', innerBlocks: [] },
 				],
 			},
-		] );
-	} );
-} );
-
-describe( 'getLastInsertedBlocksClientIds', () => {
-	it( 'should return undefined if no blocks have been inserted', () => {
-		const state = {
-			lastBlockInserted: {},
-		};
-
-		expect( getLastInsertedBlocksClientIds( state ) ).toEqual( undefined );
-	} );
-
-	it( 'should return clientIds if blocks have been inserted', () => {
-		const state = {
-			lastBlockInserted: {
-				clientIds: [ '123456', '78910' ],
-			},
-		};
-
-		expect( getLastInsertedBlocksClientIds( state ) ).toEqual( [
-			'123456',
-			'78910',
 		] );
 	} );
 } );
